# Optimize performance of GetHashCode for ContentType comparers, remove allocs