--- a/src/Microsoft.DotNet.Wpf/src/WindowsBase/MS/Internal/ContentType.cs
+++ b/src/Microsoft.DotNet.Wpf/src/WindowsBase/MS/Internal/ContentType.cs
@@ -337,13 +337,11 @@ namespace MS.Internal
             }            
 
             /// <summary>
-            /// We lower case the results of ToString() because it returns the original
-            /// casing passed into the constructor.  ContentTypes that are equal (which
-            /// ignores casing) must have the same hash code.
+            /// ContentTypes that are equal (which ignores casing) must have the same hash code.
             /// </summary>
             public int GetHashCode(ContentType obj)
             {
-                return obj.ToString().ToUpperInvariant().GetHashCode();
+                return StringComparer.OrdinalIgnoreCase.GetHashCode(obj.ToString());
             }
         }
 
@@ -366,13 +364,11 @@ namespace MS.Internal
             }
 
             /// <summary>
-            /// We lower case the results of ToString() because it returns the original
-            /// casing passed into the constructor.  ContentTypes that are equal (which
-            /// ignores casing) must have the same hash code.
+            /// ContentTypes that are equal (which ignores casing) must have the same hash code.
             /// </summary>
             public int GetHashCode(ContentType obj)
             {
-                return obj._type.ToUpperInvariant().GetHashCode() ^ obj._subType.ToUpperInvariant().GetHashCode();
+                return StringComparer.OrdinalIgnoreCase.GetHashCode(obj._type) ^ StringComparer.OrdinalIgnoreCase.GetHashCode(obj._subType);
             }
         }
         #endregion Nested Classes