## Description

Removes allocations for `ContentType`'s `WeakComparer` and `StrongComparer` (which are used in the related dictionaries).
- We will use `StringComparer.OrdinalIgnoreCase.GetHashCode(src)` instead of `src.ToUpperInvariant.GetHashCode()` as that creates a new string everytime just so we can retrieve HashCode out of it.
- This will reduce allocations to zero and also save about 40ns (depending on length, it scales up) per invocation while preserving functionality.

### Benchmarks

| Method  | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  34.736 ns |  0.2199 ns |   0.2057 ns | 0.0038 |         392 B |          64 B |
| PR_EDIT  | 8.620 ns |  0.0279 ns |   0.0247 ns |      - |         330 B |             - |

| Method   | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  70.84 ns |   1.404 ns |    1.379 ns | 0.0110 |         394 B |         184 B |
| PR_EDIT  |  41.27 ns |   0.127 ns |    0.113 ns |      - |         316 B |             - |

| Method   | Mean [ns] | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|--------- |----------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original |  183.5 ns |    2.76 ns |     2.58 ns | 0.0310 |         394 B |         520 B |
| PR_EDIT  | 132.4 ns |    0.40 ns |     0.35 ns |      - |         333 B |             - |

### Benchmark code

```csharp
//First invocation
return "aBCD".ToUpperInvariant().GetHashCode() ^ "xDDAH".ToUpperInvariant().GetHashCode();
return StringComparer.OrdinalIgnoreCase.GetHashCode("aBCD") ^ StringComparer.OrdinalIgnoreCase.GetHashCode("xDDAH");

//Second invocation
return "aBCDnowimagineifthisisLongstring".ToUpperInvariant().GetHashCode() ^ "xDDAHaBCDnowimagineifthisisLongstring".ToUpperInvariant().GetHashCode();
return StringComparer.OrdinalIgnoreCase.GetHashCode("aBCDnowimagineifthisisLongstring") ^ StringComparer.OrdinalIgnoreCase.GetHashCode("xDDAHaBCDnowimagineifthisisLongstring");

//Third invocation
return "aBCDnowimagineifthisisLongstringWeNeedthistobewaylongeeeeeeerrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr".ToUpperInvariant().GetHashCode() ^
"xDDAHaBCDnowimagineifthisisLongstringaBCDnowimagineifthisisLongstringWeNeedthistobewaylongeeeeeeerrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr".ToUpperInvariant().GetHashCode();
return StringComparer.OrdinalIgnoreCase.GetHashCode("aBCDnowimagineifthisisLongstringWeNeedthistobewaylongeeeeeeerrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr") ^ 
StringComparer.OrdinalIgnoreCase.GetHashCode("xDDAHaBCDnowimagineifthisisLongstringaBCDnowimagineifthisisLongstringWeNeedthistobewaylongeeeeeeerrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr");
```

## Customer Impact

Increased performance, including start-up performance, decreased allocations.

## Regression

No.

## Testing

Local build, assert test:

```csharp
int hashCode1 = "áBCD".ToUpperInvariant().GetHashCode() ^ "xDDAH".ToUpperInvariant().GetHashCode();
int hashCode2 = StringComparer.OrdinalIgnoreCase.GetHashCode("ÁBCD") ^ StringComparer.OrdinalIgnoreCase.GetHashCode("xDDAH");
Assert.AreEqual(hashCode1, hashCode2);
```

## Risk

Low.