@h3xds1nz 2024-09-05
CI failure caused by #9632 requesting `System.Drawing.Common` from .NET 8 (v8.0.8).